pyproject.toml
[tool.poetry]
name = "electronics-retail"
version = "0.1.0"
description = ""
authors = ["Nikita <example@mail>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.6"
dotenv = "^0.9.9"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.16.1"
black = "^25.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
